+++ src/main.js
@@ -2,5 +2,6 @@ import { createApp } from 'vue'
 import './style.css'
 import App from './App.vue'
 import store from './store'
+import router from './router'
 
-createApp(App).use(store).mount('#app')
+createApp(App).use(store).use(router).mount('#app')
