+++ src/style.css
@@ -26,7 +26,7 @@
   }
 
   .nav-pill {
-    @apply relative rounded-full px-4 py-2 text-sm font-medium transition-colors duration-300 ease-out hover:text-primary;
+    @apply relative rounded-full px-4 py-2 text-sm font-medium transition-colors duration-300 ease-out;
   }
 
   .nav-pill-active {
